third_party/blink/web_tests/fast/css/zoom-media-queries-expected.txt

PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false


testRunner.zoomPageIn()
PASS matchMedia("(max-width: 300px)").matches is matchMedia("(max-width: 18.75em)").matches
testRunner.zoomPageIn()
PASS matchMedia("(max-width: 300px)").matches is matchMedia("(max-width: 18.75em)").matches
testRunner.zoomPageIn()
PASS matchMedia("(max-width: 300px)").matches is matchMedia("(max-width: 18.75em)").matches
testRunner.zoomPageIn()
PASS matchMedia("(max-width: 300px)").matches is matchMedia("(max-width: 18.75em)").matches
testRunner.zoomPageIn()
PASS matchMedia("(max-width: 300px)").matches is matchMedia("(max-width: 18.75em)").matches
testRunner.zoomPageIn()


PASS matchMedia("(max-width: 300px)").matches is true
PASS matchMedia("(max-width: 18.75em)").matches is true


testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
testRunner.zoomPageOut()
PASS matchMedia("(max-width: 300px)").matches is false
PASS matchMedia("(max-width: 18.75em)").matches is false
PASS successfullyParsed is true

TEST COMPLETE
